bugfix: added crontab script

+++ scredits-crontab-script.sh
@@ -6,7 +6,7 @@
 #              This script prunes Slurm accounts and users raw usage after N months, resettings billing and resouces consumpion (cpu, gpu, mem). 
 #              It supports verbose mode, specifying multiple clusters, and sets the next prune date.
 # Author:      Giulio Librando <giuliolibrando@gmail.com>
-# Created:     09/07/2024 v1
+# Created:     09/07/2024 v1.1
 # License:     MIT
 # =============================================================================
 
@@ -88,14 +88,14 @@ fi
 # Ensure file /etc/scredits/last_prune exists and has correct permissions
 LAST_PRUNE_FILE="$SCREDITS_DIR/last_prune"
 if [[ ! -f "$LAST_PRUNE_FILE" ]]; then
-    echo "2024-01-01-00-00" > "$LAST_PRUNE_FILE"
+    echo "2024-07-09-15-47" > "$LAST_PRUNE_FILE"
     chmod 644 "$LAST_PRUNE_FILE"
 fi
 
 # Ensure file /etc/scredits/next_prune exists and has correct permissions
 NEXT_PRUNE_FILE="$SCREDITS_DIR/next_prune"
 if [[ ! -f "$NEXT_PRUNE_FILE" ]]; then
-    echo "2024-01-01-00-00" > "$NEXT_PRUNE_FILE"
+    echo "2024-10-31-23-59" > "$NEXT_PRUNE_FILE"
     chmod 644 "$NEXT_PRUNE_FILE"
 fi
 
@@ -112,6 +112,17 @@ fi
 # Execute "sacctmgr list account -P" and save account names in ACCOUNT_LIST variable
 ACCOUNT_LIST=$(sacctmgr list account -P | awk -F'|' '{print $1}')
 
+# Estrai anno, mese, giorno, ora e minuti dalla data
+YEAR=$(echo "$SCREDITS_NEXT_PRUNE" | cut -d'-' -f1)
+MONTH=$(echo "$SCREDITS_NEXT_PRUNE" | cut -d'-' -f2)
+DAY=$(echo "$SCREDITS_NEXT_PRUNE" | cut -d'-' -f3)
+HOUR=$(echo "$SCREDITS_NEXT_PRUNE" | cut -d'-' -f4)
+MINUTE=$(echo "$SCREDITS_NEXT_PRUNE" | cut -d'-' -f5)
+
+# Costruisci la data in formato Unix timestamp
+SCREDITS_NEXT_PRUNE_TIMESTAMP=$(date -d "${YEAR}-${MONTH}-${DAY} ${HOUR}:${MINUTE}:00" +'%s')
+
+echo "Timestamp Unix per SCREDITS_NEXT_PRUNE: $NEXT_PRUNE_TIMESTAMP"
 # Loop over each account retrieved
 for account in $ACCOUNT_LIST; do
     # Skip the header "Account"
@@ -119,42 +130,46 @@ for account in $ACCOUNT_LIST; do
         continue
     fi
 
-    verbose_echo "Modifying account $account"
+    # Get current timestamp
+    CURRENT_TIMESTAMP=$(date +'%s')
 
-    # Create temporary file for 'y' input
-    TMPFILE=$(mktemp)
-    echo "y" > "$TMPFILE"
+    # Compare current timestamp with SCREDITS_NEXT_PRUNE timestamp
+    if [[ "$CURRENT_TIMESTAMP" -ge "$SCREDITS_NEXT_PRUNE_TIMESTAMP" ]]; then
+        verbose_echo "Modifying account $account"
 
-    # Execute command to modify account for each cluster, feeding 'y' from TMPFILE
-    if [[ "$CLUSTER_SPECIFIED" == "true" ]]; then
-        IFS=',' read -r -a CLUSTERS <<< "$CLUSTER"
-        for cl in "${CLUSTERS[@]}"; do
-            verbose_echo "  in cluster $cl"
-            sacctmgr modify account $account cluster="$cl" set RawUsage= < "$TMPFILE"
-        done
+        # Create temporary file for 'y' input
+        TMPFILE=$(mktemp)
+        echo "y" > "$TMPFILE"
+
+        # Execute command to modify account for each cluster, feeding 'y' from TMPFILE
+        if [[ "$CLUSTER_SPECIFIED" == "true" ]]; then
+            IFS=',' read -r -a CLUSTERS <<< "$CLUSTER"
+            for cl in "${CLUSTERS[@]}"; do
+                verbose_echo "  in cluster $cl"
+                sacctmgr modify account $account cluster="$cl" set RawUsage= < "$TMPFILE"
+            done
+        else
+            sacctmgr modify account $account set RawUsage= < "$TMPFILE"
+        fi
+
+        # Remove temporary file
+        rm "$TMPFILE"
     else
-        sacctmgr modify account $account set RawUsage= < "$TMPFILE"
+        verbose_echo "Skipping modification of account $account because current date is not yet at SCREDITS_NEXT_PRUNE or has passed"
     fi
-
-    # Remove temporary file
-    rm "$TMPFILE"
 done
-
 # Set SCREDITS_LAST_PRUNE to current date and time
 SCREDITS_LAST_PRUNE=$(date +'%Y-%m-%d-%H-%M')
 
 # Calculate the date N months ahead from the first day of the current month
-NEXT_MONTH=$(date -d "$(date +'%Y-%m-01') +1 month")
+CURRENT_MONTH=$(date +'%Y-%m')
+NEXT_PRUNE=$(date -d "$CURRENT_MONTH-01 +$((MONTHS-1)) months" +'%Y-%m-%d')
 
-# Calculate the last day of the month for the NEXT_MONTH
-LAST_DAY=$(date -d "$NEXT_MONTH -1 day")
-
-# Calculate the final NEXT_PRUNE datetime
-NEXT_PRUNE=$(date -d "$LAST_DAY 23:59 +$((MONTHS-1)) months")
+# Calculate the last day of the month for the NEXT_PRUNE
+LAST_DAY=$(date -d "$NEXT_PRUNE +1 month -1 day" +'%Y-%m-%d')
 
 # Format SCREDITS_NEXT_PRUNE as yyyy-mm-dd-hh-mm
-SCREDITS_NEXT_PRUNE=$(date -d "$(date +'%Y-%m-01') +$((MONTHS+1)) months -1 day 23:59")
-SCREDITS_NEXT_PRUNE=$(date -d "$SCREDITS_NEXT_PRUNE" +'%Y-%m-%d-%H-%M')
+SCREDITS_NEXT_PRUNE=$(date -d "$LAST_DAY 23:59" +'%Y-%m-%d-%H-%M')
 
 # Update /etc/scredits/last_prune and /etc/scredits/next_prune with new values
 echo "$SCREDITS_LAST_PRUNE" > "$LAST_PRUNE_FILE"
